rename "find" route to "getById"
