+++ x-pack/plugins/files/server/routes/file_kind/get_by_id.ts
